Tests: Move block raw handling tests to integration folder
